[ghstack-poisoned]

+++ torchtitan/models/qwen3/model/args.py
@@ -59,7 +59,5 @@ class Qwen3ModelArgs(BaseModelArgs):
             job_config.debug.moe_force_load_balance
         )
 
-    def get_nparams_and_flops(
-        self, model: nn.Module, seq_len: int
-    ) -> tuple[int, float]:
+    def get_nparams_and_flops(self, model: nn.Module, seq_len: int) -> tuple[int, int]:
         return get_moe_model_nparams_and_flops(self, model, 2 * self.head_dim, seq_len)
